Task_B/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
timm>=0.6.12
opencv-python>=4.6.0
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
albumentations>=1.3.0
pathlib2>=2.3.7
json5>=0.9.0
logging>=0.4.9.6
